+++ CHANGELOG.md
@@ -7,7 +7,7 @@ PrestaShop 1.6.x, 1.7.x and 8.x
 
 ## [3.8.1]
 ### Added
-- Support "Pay by Link" functionality with the ability to edit existing orders by replacing or increasing the quantity.
+- Generate a payment link to capture the additional amount after the order update.
 
 ## [3.8.0]
 ### Added